% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-string.R
\name{.anchor_regex}
\alias{.anchor_regex}
\title{Operator: matches_regex}
\usage{
.anchor_regex(pat)
}
\description{
Check that column values match a PCRE pattern. NA / empty are
considered compliant unless allow_missing = FALSE.
}
