% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-detect.R
\name{.expand_col_pattern}
\alias{.expand_col_pattern}
\title{Expand a column pattern: \verb{--<STEM>} becomes \verb{<DOM2><STEM>} using the
first two chars of \code{ds_name}; otherwise the pattern is returned as-is.}
\usage{
.expand_col_pattern(pat, ds_name)
}
\description{
Expand a column pattern: \verb{--<STEM>} becomes \verb{<DOM2><STEM>} using the
first two chars of \code{ds_name}; otherwise the pattern is returned as-is.
}
