% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict-provider.R
\name{print.herald_dict_provider}
\alias{print.herald_dict_provider}
\title{Print a herald_dict_provider}
\usage{
\method{print}{herald_dict_provider}(x, ...)
}
\arguments{
\item{x}{A \code{herald_dict_provider} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a herald_dict_provider
}
