% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val-result.R
\name{print.herald_result}
\alias{print.herald_result}
\title{Print a herald_result}
\usage{
\method{print}{herald_result}(x, ...)
}
\arguments{
\item{x}{A \code{herald_result} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a herald_result
}
