% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-read.R
\name{print.herald_spec}
\alias{print.herald_spec}
\title{Print a herald_spec}
\usage{
\method{print}{herald_spec}(x, ...)
}
\arguments{
\item{x}{A \code{herald_spec} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a herald_spec
}
