% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-read.R
\name{read_define_xml}
\alias{read_define_xml}
\title{Read a Define-XML 2.1 file}
\usage{
read_define_xml(path, call = rlang::caller_env())
}
\arguments{
\item{path}{Path to the Define-XML file.}

\item{call}{Caller environment for error reporting.}
}
\value{
A \code{herald_define} S3 object (inherits from \code{list}).
}
\description{
Parses a Define-XML 2.1 file and returns a \code{herald_define} object
carrying datasets, variables, codelists, methods, comments, ARM metadata,
sponsor-defined key variables, and the LOINC dictionary version declared
in \code{MetaDataVersion}. Requires the \code{xml2} package.
}
\examples{
if (requireNamespace("xml2", quietly = TRUE)) {
  tmp <- tempfile(fileext = ".xml")
  on.exit(unlink(tmp))
  writeLines(c(
    '<?xml version="1.0" encoding="UTF-8"?>',
    '<ODM xmlns="http://www.cdisc.org/ns/odm/v1.3"',
    '     xmlns:def="http://www.cdisc.org/ns/def/v2.1">',
    '  <Study OID="S.TEST">',
    '    <GlobalVariables>',
    '      <StudyName>PILOT01</StudyName>',
    '    </GlobalVariables>',
    '    <MetaDataVersion OID="MDV.1" Name="MDV1">',
    '      <ItemGroupDef OID="IG.DM" Name="DM" Repeating="No">',
    '        <Description>',
    '          <TranslatedText>Demographics</TranslatedText>',
    '        </Description>',
    '      </ItemGroupDef>',
    '    </MetaDataVersion>',
    '  </Study>',
    '</ODM>'
  ), tmp)
  d <- read_define_xml(tmp)
  d$ds_spec
}

}
\seealso{
Other spec: 
\code{\link{apply_spec}()},
\code{\link{as_herald_spec}()},
\code{\link{herald_spec}()},
\code{\link{is_herald_spec}()},
\code{\link{write_define_xml}()}
}
\concept{spec}
