% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herald-package.R
\docType{package}
\name{herald-package}
\alias{herald}
\alias{herald-package}
\title{herald: CDISC Conformance Validation for Clinical Trial Submissions}
\description{
Pure-R conformance validation for CDISC clinical data submissions.
Checks SDTM and ADaM datasets against published CDISC Library conformance
rules, controlled terminology, and Define-XML 2.1 specifications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/about-vignesh/herald}
  \item \url{https://about-vignesh.github.io/herald/}
  \item Report bugs at \url{https://github.com/about-vignesh/herald/issues}
}

}
\author{
\strong{Maintainer}: Vignesh \email{about.vignesh@gmail.com}

}
\keyword{internal}
