% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-read.R
\name{is_herald_spec}
\alias{is_herald_spec}
\title{Is \code{x} a \code{herald_spec}?}
\usage{
is_herald_spec(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{herald_spec}, else \code{FALSE}.
}
\description{
Is \code{x} a \code{herald_spec}?
}
\seealso{
Other spec: 
\code{\link{apply_spec}()},
\code{\link{as_herald_spec}()},
\code{\link{herald_spec}()},
\code{\link{read_define_xml}()},
\code{\link{write_define_html}()},
\code{\link{write_define_xml}()}
}
\concept{spec}
