% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict-providers-ext.R
\name{loinc_provider}
\alias{loinc_provider}
\title{LOINC as a Dictionary Provider}
\usage{
loinc_provider(path, version = "unknown")
}
\arguments{
\item{path}{Path to \code{Loinc.csv} OR a directory containing it.}

\item{version}{LOINC release tag (e.g. \code{"2.77"}). User-supplied.}
}
\value{
A \code{herald_dict_provider} with name \code{"loinc"}.
}
\description{
Reads a Regenstrief LOINC CSV distribution (\code{Loinc.csv}) and
returns a \code{herald_dict_provider} with membership lookup by
\code{LOINC_NUM}, \code{COMPONENT}, \code{SHORTNAME}, or \code{LONG_COMMON_NAME}.
}
\seealso{
\code{\link[=register_dictionary]{register_dictionary()}}.

Other dict: 
\code{\link{ct_provider}()},
\code{\link{custom_provider}()},
\code{\link{list_dictionaries}()},
\code{\link{meddra_provider}()},
\code{\link{new_dict_provider}()},
\code{\link{register_dictionary}()},
\code{\link{snomed_provider}()},
\code{\link{srs_provider}()},
\code{\link{unregister_dictionary}()},
\code{\link{whodrug_provider}()}
}
\concept{dict}
