% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-read.R
\name{print.herald_define}
\alias{print.herald_define}
\title{Print a herald_define}
\usage{
\method{print}{herald_define}(x, ...)
}
\arguments{
\item{x}{A \code{herald_define} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a herald_define
}
\seealso{
Other methods: 
\code{\link{herald_submission-methods}},
\code{\link{print.herald_dict_provider}()},
\code{\link{print.herald_result}()},
\code{\link{print.herald_spec}()},
\code{\link{summary.herald_result}()}
}
\concept{methods}
