% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val-result.R
\name{print.herald_result}
\alias{print.herald_result}
\title{Print a herald_result}
\usage{
\method{print}{herald_result}(x, ...)
}
\arguments{
\item{x}{A \code{herald_result} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a herald_result
}
\seealso{
Other methods: 
\code{\link{herald_submission-methods}},
\code{\link{print.herald_define}()},
\code{\link{print.herald_dict_provider}()},
\code{\link{print.herald_spec}()},
\code{\link{summary.herald_result}()}
}
\concept{methods}
