% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-xlsx.R
\name{write_report_xlsx}
\alias{write_report_xlsx}
\title{Write a herald_result as an XLSX workbook}
\usage{
write_report_xlsx(x, path, ...)
}
\arguments{
\item{x}{A \code{herald_result} object returned by \code{\link[=validate]{validate()}}.}

\item{path}{Output file path (should end in \code{.xlsx}).}

\item{...}{Ignored.}
}
\value{
\code{path}, invisibly.
}
\description{
Write a herald_result as an XLSX workbook
}
\seealso{
Other report: 
\code{\link{report}()},
\code{\link{write_report_html}()},
\code{\link{write_report_json}()}
}
\concept{report}
